requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
msgraph-sdk
aiohttp